+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Gluetex bugfixes
+_25.06.2026_
+
+- Improved handling after a safety stop that shouldnt need the workaround with standby -> Wind, which cleared the traverse position
+
 # `Small Gluetex Updates`
 _15.06.2026_